requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-dump-die==0.1.5
djangorestframework==3.14.0
idna==3.4
mysqlclient==2.1.1
Pillow==9.5.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.0
tzdata==2023.3
urllib3==2.0.3
